adding testing, closes #16, #21
